Final Push

+++ CFarmacia.hpp
@@ -28,7 +28,7 @@ class CFarmacia{
     CFarmacia(int id, CPos* pos,list<CMedFarm*> medicamentos);
     ~CFarmacia();
     CCompra* determine_compra();//returns what it needs to buy(quantity) but not from whom or at what price.
-    CVenta* make_venta(int id_med,int cantidad_venta);
+    CVenta* make_venta(float id_med,int cantidad_venta);
     int get_ID();
     friend std::ostream& operator<<(std::ostream& os, CFarmacia* farmacia);
