+++ CMedicamento.hpp
@@ -8,14 +8,14 @@
 
 	class CMedicamento{
 	private:
-		int m_i_cn;
+		float m_i_cn;
 	public:
 		CMedicamento(void);
 		CMedicamento(const CMedicamento& medicamento);
-		CMedicamento(int m_i_cn);
+		CMedicamento(float m_i_cn);
 		~CMedicamento();
 
-		int get_cn(void);
+		float get_cn(void);
 
 		friend std::ostream& operator<<(std::ostream& os, const CMedicamento& medicamento);
 		friend std::istream& operator>>(std::istream& i, CMedicamento& medicamento);
